requirements.txt
# Core Dependencies
streamlit==1.28.1

# Data Processing
pandas==2.1.0
numpy==1.24.0

# Machine Learning & Datasets
scikit-learn==1.3.0

# Visualization
plotly==5.17.0

# Utilities
python-dotenv==1.0.0

